\name{PBRF-package}
\alias{PBRF-package}
\alias{pbrf}
\docType{package}
\title{
\packageTitle{PBRF}
}
\description{
\packageDescription{PBRF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PBRF}
\packageIndices{PBRF}
}
\author{
\packageAuthor{PBRF}

Maintainer: \packageMaintainer{PBRF}
}
\references{
Tsai W.Y., Luo X., Crowley J. (2017) The Probability of Being in Response Function and Its Applications. In: Matsui S., Crowley J. (eds) Frontiers of Biostatistical Methods and Applications in Clinical Oncology. Springer, Singapore. <doi: https://doi.org/10.1007/978-981-10-0126-0_10>.
}
\keyword{survival analysis}
\keyword{composite endpoints}
\keyword{multiple events}

